PRO-5159: Made expiryDate readonly - removed validation and tests
