+++ ccdImports/configFiles/CCD_Probate_Caveat/AuthorisationCaseField.json
@@ -22,7 +22,7 @@
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorSurname", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorEmailAddress", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorAddress", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
-  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
+  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-issuer", "CRUD": "CRU"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "messageContent", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatReopenReason", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "documentsUploaded", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
@@ -67,7 +67,7 @@
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorSurname", "UserRole": "caseworker-probate-caseadmin", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorEmailAddress", "UserRole": "caseworker-probate-caseadmin", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorAddress", "UserRole": "caseworker-probate-caseadmin", "CRUD": "CRUD"},
-  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-caseadmin", "CRUD": "CRUD"},
+  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-caseadmin", "CRUD": "CRU"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "messageContent", "UserRole": "caseworker-probate-caseadmin", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatReopenReason", "UserRole": "caseworker-probate-caseadmin", "CRUD": "R"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "documentsUploaded", "UserRole": "caseworker-probate-caseadmin", "CRUD": "CRUD"},
@@ -111,7 +111,7 @@
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorSurname", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorEmailAddress", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorAddress", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
-  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
+  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRU"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "messageContent", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatReopenReason", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "documentsUploaded", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
@@ -154,7 +154,7 @@
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorSurname", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorEmailAddress", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorAddress", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
-  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
+  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-registrar", "CRUD": "CRU"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "messageContent", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatReopenReason", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "documentsUploaded", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
@@ -197,7 +197,7 @@
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorSurname", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorEmailAddress", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatorAddress", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
-  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
+  {"CaseTypeID": "Caveat", "CaseFieldID": "expiryDate", "UserRole": "caseworker-probate-superuser", "CRUD": "CRU"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "messageContent", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "caveatReopenReason", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseFieldID": "documentsUploaded", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
