+++ ccdImports/configFiles/CCD_Probate_Caveat/CaseEvent.json
@@ -27,7 +27,7 @@
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "ID": "withdrawCaveatForAwaitingCaveatResolution", "Name": "Withdraw caveat", "Description": "Withdraw caveat", "DisplayOrder": 26, "PreConditionState(s)": "AwaitingCaveatResolution", "PostConditionState": "CaveatClosed", "SecurityClassification": "Public", "ShowSummary": "N"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "ID": "caveatExpiredForAwaitingCaveatResolution", "Name": "Caveat expired", "Description": "Caveat expired", "DisplayOrder": 27, "PreConditionState(s)": "AwaitingCaveatResolution", "PostConditionState": "CaveatClosed", "SecurityClassification": "Public", "ShowSummary": "N"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "ID": "warningRequested", "Name": "Warning requested", "Description": "Warning requested", "DisplayOrder": 28, "PreConditionState(s)": "AwaitingCaveatResolution", "PostConditionState": "WarningValidation", "SecurityClassification": "Public", "ShowSummary": "N"},
-  {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "ID": "extendCaveat", "Name": "Extend Caveat", "Description": "Extend Caveat", "DisplayOrder": 28, "PreConditionState(s)": "AwaitingCaveatResolution", "PostConditionState": "AwaitingCaveatResolution", "CallBackURLAboutToStartEvent": "http://${CCD_DEF_CASE_SERVICE_BASE_URL}/caveat/validate-extend", "CallBackURLAboutToSubmitEvent": "http://${CCD_DEF_CASE_SERVICE_BASE_URL}/caveat/extend", "SecurityClassification": "Public", "ShowSummary": "N"},
+  {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "ID": "extendCaveat", "Name": "Extend Caveat", "Description": "Extend Caveat", "DisplayOrder": 28, "PreConditionState(s)": "AwaitingCaveatResolution", "PostConditionState": "AwaitingCaveatResolution", "CallBackURLAboutToStartEvent": "http://${CCD_DEF_CASE_SERVICE_BASE_URL}/caveat/validate-extend", "SecurityClassification": "Public", "ShowSummary": "N"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "ID": "amendCaveatForWarningValidation", "Name": "Amend caveat details", "Description": "Amend caveat details", "DisplayOrder": 29, "PreConditionState(s)": "WarningValidation", "PostConditionState": "WarningValidation", "SecurityClassification": "Public", "ShowSummary": "N"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "ID": "uploadDocumentForWarningValidation", "Name": "Upload document", "Description": "Upload document", "DisplayOrder": 30, "PreConditionState(s)": "WarningValidation", "PostConditionState": "WarningValidation", "SecurityClassification": "Public", "ShowSummary": "N"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "ID": "emailCaveatorForWarningValidation", "Name": "Email caveator", "Description": "Email caveator", "DisplayOrder": 31, "PreConditionState(s)": "WarningValidation", "PostConditionState": "WarningValidation", "CallBackURLAboutToSubmitEvent": "http://${CCD_DEF_CASE_SERVICE_BASE_URL}/caveat/general-message", "SecurityClassification": "Public", "ShowSummary": "N"},
