+++ ccdImports/configFiles/CCD_Probate_Caveat/CaseEventToFields.json
@@ -57,7 +57,7 @@
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "CaseEventID": "amendCaveatForAwaitingCaveatResolution", "CaseFieldID": "caveatorAddress", "PageFieldDisplayOrder": 5, "DisplayContext": "MANDATORY", "PageID": "amendCaveatDetailsPage3", "PageLabel": "Caveator details", "PageDisplayOrder": 3, "ShowSummaryChangeOption": "y"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "CaseEventID": "uploadDocumentForAwaitingCaveatResolution", "CaseFieldID": "documentsUploaded", "PageFieldDisplayOrder": 1, "DisplayContext": "MANDATORY", "PageID": "uploadDocumentPage1", "PageLabel": "Document upload", "PageDisplayOrder": 1, "ShowSummaryChangeOption": "n"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "CaseEventID": "emailCaveatorForAwaitingCaveatResolution", "CaseFieldID": "messageContent", "PageFieldDisplayOrder": 1, "DisplayContext": "MANDATORY", "PageID": "emailCaveatorPage1", "PageLabel": "Email caveator", "PageDisplayOrder": 1, "ShowSummaryChangeOption": "n"},
-  {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "CaseEventID": "extendCaveat", "CaseFieldID": "expiryDate", "PageFieldDisplayOrder": 1, "DisplayContext": "MANDATORY", "PageID": "extendCaveatPage1", "PageLabel": "Extend caveat expiry", "PageDisplayOrder": 1, "ShowSummaryChangeOption": "n"},
+  {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "CaseEventID": "extendCaveat", "CaseFieldID": "expiryDate", "PageFieldDisplayOrder": 1, "DisplayContext": "READONLY", "PageID": "extendCaveatPage1", "PageLabel": "Extend caveat expiry", "PageDisplayOrder": 1, "ShowSummaryChangeOption": "n"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "CaseEventID": "amendCaveatForWarningValidation", "CaseFieldID": "applicationType", "PageFieldDisplayOrder": 1, "DisplayContext": "MANDATORY", "PageID": "amendCaveatDetailsPage1", "PageLabel": "Case details", "PageDisplayOrder": 1, "ShowSummaryChangeOption": "y"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "CaseEventID": "amendCaveatForWarningValidation", "CaseFieldID": "registryLocation", "PageFieldDisplayOrder": 2, "DisplayContext": "MANDATORY", "PageID": "amendCaveatDetailsPage1", "PageLabel": "Case details", "PageDisplayOrder": 1, "ShowSummaryChangeOption": "y"},
   {"LiveFrom": "01/01/2017", "CaseTypeID": "Caveat", "CaseEventID": "amendCaveatForWarningValidation", "CaseFieldID": "paperForm", "PageFieldDisplayOrder": 3, "DisplayContext": "MANDATORY", "PageID": "amendCaveatDetailsPage1", "PageLabel": "Case details", "PageDisplayOrder": 1, "ShowSummaryChangeOption": "y"},
