+++ src/main/java/uk/gov/hmcts/probate/validator/CaveatsExpiryValidationRule.java
@@ -1,19 +1,4 @@
 package uk.gov.hmcts.probate.validator;
 
-import com.google.common.base.Strings;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.probate.exception.model.FieldErrorResponse;
-import uk.gov.hmcts.probate.model.ccd.CCDData;
-import uk.gov.hmcts.probate.model.ccd.raw.SolsAddress;
-import uk.gov.hmcts.probate.service.BusinessValidationMessageService;
-
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-import static uk.gov.hmcts.probate.model.Constants.BUSINESS_ERROR;
-
 public interface CaveatsExpiryValidationRule extends ValidationRuleCaveats {
 }
